refactor: useDragDrop step 2 — drag/hover state and core refs
Moves 10 useState and 3 useRef declarations out of App.jsx into
useDragDrop: draggedTask, dragSource, dragPreviewTime, dragPreviewDate,
dragOverAllDay, dragOverInbox, dragOverRecycleBin, hoverPreviewTime,
hoverPreviewDate, isResizing, autoScrollInterval, frameResizingRef,
stickyHeaderRef. All destructured back into App.jsx — no behaviour change.
